Torch.Core.Random

Description

Random number generation for single values. FFI over TH/THRandom.h

Documentation

data Generator

Representation of a CPU-bound random number generator

data Seed

Representation of a CPU-bound random seed

newRNG :: IO Generator

Construct a new Generator

copy :: Generator -> Generator -> IO Generator

Copy a Generator state to a new generator

seed :: Generator -> IO Seed

Get the current Seed of a Generator

manualSeed :: Generator -> Seed -> IO ()

Manually set the seed Seed of a Generator

initialSeed :: Generator -> IO Seed

Get the first Seed that initialized a given Generator

random :: Generator -> IO Seed

random64 :: Generator -> IO Seed

uniform

Arguments

:: Generator

-> Double
lower bound

-> Double
upper bound

-> IO Double

Returns a random double according to uniform distribution on [a,b).

uniformFloat

Arguments

:: Generator

-> Float
lower bound

-> Float
upper bound

-> IO Float

Returns a random float according to uniform distribution on [a,b).

normal

Arguments

:: Generator

-> Double
mean

-> Double
stddev (must be positive)

-> IO Double

Returns a random real number according to a normal distribution with the given mean and standard deviation stdv. stdv must be positive, but this is not yet enforced.

TODO: add a newtype Pos a = Pos { getPos :: a } package with a smart constructor export

exponential :: Generator -> Double -> IO Double

Returns a random real number according to the exponential distribution p(x) = lambda * exp(-lambda * x)

standard_gamma :: Generator -> Double -> IO Double

cauchy

Arguments

:: Generator

-> Double
median

-> Double
sigma

-> IO Double

Returns a random real number according to the Cauchy distribution p(x) = sigma/(pi*(sigma^2 + (x-median)^2))

logNormal

Arguments

:: Generator

-> Double
mean

-> Double
stddev (must be positive)

-> IO Double

Returns a random real number according to the log-normal distribution, with the given mean and standard deviation stdv. mean and stdv are the corresponding mean and standard deviation of the underlying normal distribution, and not of the returned distribution.

stdv must be positive.

geometric :: Generator -> Double -> IO Int

Returns a random integer number according to a geometric distribution p(i) = (1-p) * p^(i-1). p must satisfy 0 < p < 1.

bernoulli :: Generator -> Double -> IO Int

Returns 1 with probability p and 0 with probability 1-p. p must satisfy 0 <= p <= 1.

TODO: By default p is equal to 0.5 -- this isn't encoded in the API